requirements.txt
asyncio==3.4.3
certifi==2018.4.16
chardet==3.0.4
dotenv-python==0.0.1
elasticsearch==6.2.0
idna==2.6
python-dotenv==0.8.2
requests==2.18.4
urllib3==1.22
websockets==5.0
